(PE-15351) Fix use of -c/-f flags on upgrades
Prior to this commit the condition used to decide whether to use
the `-c`/`-f` flags was dependent on `host['pe_upgrade_ver']` and
`host['pe_ver']` which was an unreliable condition.
This commit updates the condition to determine whether to use the
`-f` flag to simply look at `host['pe_ver']` because that value
is updated depending on what version of pe is currently being
installed or upgraded to.
The condition to decide to omit the `-c` flag has to depend on
`opts[:HOSTS][host.name][:pe_ver]` because that value is not
modified during upgrade and can be used for a valid comparison
to determine if the install will have a `pe.conf` file to use
for an upgrade.