(PE-16566) Add method to download life support gpg key
As of July 8th, 2016 the GPG key that was shipped with and used to sign
repos inside of PE tarballs expired. A new life support key was created
that extended the expiration date to Jan 2017. That key shipped with PE
3.8.5 and 2016.1.2.

apt based platforms appear to be the only package manager failing due to
an expired key, while rpm is fine.

This commit adds a new helper method to additionally download and
install the extended key for PE versions that have already been released
and are needing to be tested.