(PE-16886) Add wait for console to be functional
Before this commit the console may or may not be functional by the time
the next puppet agent run occurs on the following node. This can cause
puppetserver to return with an error from the classifier when it is
attempting to evaluate the classes which should be applied to the node.

This commit adds in a sleep and service check to the final agent run
step on the console node which will hopefully work around this issue
until it is fixed in SERVER-1237.