(BKR-806) Only add pe_repo classes once
Previously, if installing multiple agents with the same platform, the
work to add the appropriate pe_repo class to the master would be
performed once for each agent. Because that requires running `puppet
agent -t` on the master, it can add a significant amount of unnecessary
overhead to the install process when installing several agents.

We now check whether the class has already been included in the Beaker
Frictionless Agent group and skip running puppet on the master if it
has.