(PE-19169) Add pe modules next flag (#55)
* (PE-19169,PE-18516) Remove version specific upgrade hack

This had been added to allow upgrade testing in flanders with meep for
classification prior to our having completed an upgrade workflow in the
installer.

Now that meep is being worked on in the pe-modules-next package behind a
feature-flag, this gate is counter-productive, particularly for Flanders
upgrade testing because it masks the common case of upgrading without
specifying a pe.conf.

* (PE-19169) Register flag for pe-modules-next if present in ENV

Major PE modules development work is being done in a separate
pe-modules-next package. The installer shim decides which pe-modules
package to install based on a pe.conf flag. This patch allows Jenkins CI
to trigger this by picking up ENV['PE_MODULES_NEXT'] and ensure that the
flag is set in the opts[:answers] which will then be transferred to
pe.conf by beaker-answers, and then picked up by the installer shim
during installation.

* (PE-19169) Ensure beaker-answers generates meep 2.0 schema

...if we are installing with the pe_modules_next flag.

The pe-modules-next is meep classification at the moment and needs a 2.0
pe.conf generated.

* (PE-19169) Extract a FeatureFlags class

to make it easier to test. Flags can be set in the :answers Hash or
picked up from environment variables. In the later case, we need to be
able to preserver them in the :answers hash, if not already present, so
that BeakerAnswers ensures that they are set in the generated pe.conf
file.