(PE-19049) Remove get_console_dispatcher_for_beaker_pe specs
Because scooter can in fact be in the bundle based on the GEM_SOURCE
setting, these specs can break, and it is not worth the effort reworking
the tests to be conditional based on presence or absence of the scooter
gem. RE-8616 should make the scooter gem public and then we don't need
to dance around it like this.