(BKR-1085) Sign certs *before* stopping agent service
Previously, the simple_monolithic install method was installing agents,
stopping the agent service and then attempting to sign certificates.
This was failing sporadically because the agent service needs to run
enough to submit a certificate request before being stopped. This was
being handled properly in the generic install method, but the order of
operations was swapped in the monolithic install method.