(PE-20086) PE 2016.5.(0|1) should install via msi method for windows2008r2
Due to the timing of our LTS releases and our new major branches, PE 2016.5.0 and
PE 2016.5.1 did not get the windows2008r2 powershell fix that was done in PE-18351.
This means we need to not attempt to install fricitonlessly if it is pe 2016.5.(0|1)
if the agent platform is windows2008r2.
This PR adjust the install_via_msi? method and refactors the logic in there to clean
it up a bit (it is getting tough to easily read).
It breaks the method down to three lines:
1. If the agent is older then PE 2016.4.0.
2. If the agent is windows2008r2 and is less then 2016.4.3
3. If the agent is windows2008r2 and the agent version is between 2016.4.99 and 2016.5.99.
If any of those are true then the MSI method should be used to install the agent.