(maint) Properly install pe-client-tools when using a tag version
Previously, installing pe-client-tools with a tag version would fail on
Windows/OS X and install the wrong package on Linux.

When installing pe-client-tools, we provide two options:
- pe_client_tools_sha: the commit SHA of the version to install
- pe_client_tools_version: the `git describe` of the version to install

pe_client_tools_version is always the name of the package to install.
But the *location* of the package differs based on whether the package
version corresponds to a tag or not. When the package isn't a tag
version, it's located in a directory named based on the SHA. But when it
is a tag version, it's located in a directory named after the tag.

When pe_client_tools_version was specified as a tag, we would look in
the directory named after the SHA (which was actually from a *previous*
build of the package, from before it was tagged) for a file named after
the tag. That file would never be there, since we had a mismatch of
directory and filename. For Windows and OS X, this caused a failure to
install, because they need to know the exact filename.

This case incidentally *worked* (or appeared to work) on Linux
platforms, because they never actually refer to the package by
filename. Instead, they install the package by setting up a repo config,
which *is* always named after pe_client_tools_sha, and never
pe_client_tools_version. In that case, the Linux platforms would
actually install the previous version of the package by SHA, from before
it had been tagged.

We now properly handle the case where pe_client_tools_version is a tag,
by using that version as the location of the file in addition to the
filename.