(PE-20589) After a simple monolithic install run puppet on non-infrastructure agents
Previously after agents were installed via simple mono install, we ran the agents
all at once. This include the master. Occasionaly there would be a change done to
the console services that would require a restart.
This would cause other agent runs to fail. So to not have that happen, we will run
only the non-infrastructure agents. That should be fine since part of the process
of installing PE should have the puppet runs included there.