PE-20610 Fix install failure on windows for old pe versions
Added logic to use the old generic_install method on windows when
installing old pe versions that requires an msi install because of
powershell2 issue PE-18351. The newly added simple_monolithic_install
does not have a check for those conditions and proceeds with
frictionless installation on those hosts which fails.