(QENG-5040) Run puppet after simplified mono install
Previous to this commit, the new simplified mono install method was not
running puppet on the master node in a mono only scenario. If the
install included agents not of the masters OS, then puppet would of been
ran due to needing to add pe_repo classes. In the scenario of mono
master only, if puppet is not ran, then setup is not considered complete
due to exported resources, mcollective and facts not being setup yet.
This would cause numerous issues, such as no facts in puppetdb (so
anyaltic tests for example would fail) until a test somewhere in the
pipeline ran puppet agent on the master node.
This commit adds a call outside of the parallel agent run on non infra
agents due to the fact that with exported resources, the classifier
service will restart, so need to run the master first, then after that
all the agents can run.