(QENG-5095) Windows frictionless uses current url
Previous to this commit, the url constructed for frictionless agent
installs attempted to use the hosts `pe_ver` variable for getting the
exact URL of the simplified install script. This was causing issues with
PEZ builds due to the different pe_ver in the filename (for downloading
the tarball) and the actual pe version.
This commit updates the url to just use the `current` endpoint instead
of the version, which mimics our install documentation. This won't
really remove the ability of installing an older version via this method
considering the logic for adding pe_repo classes didn't support passing
a specific pe_ver / agent version to pe_repo.