(PE-21277) Support frictionless install with ca cert
Frictionless installs are a curl | bash sequence, and because we are
curling from the primary, generally to install puppet-agent for the
first time, we don't have the primary master's ca certificate yet. So
the invocation is `curl -k` and the host's certificate is not verified.

If ca.pem certificate is copied to the host before hand, we can `curl
--cacert /etc/puppetlabs/puppet/ssl/certs/ca.pem` and validate that the
primary master certificate is valid (and that we're talking to correct
host). Changes were made to puppetlabs/puppetlabs-pe_repo in PE-21277
to propagate this secure curl through the install process.

This patch to beaker-pe allows testing this workflow with beaker by
setting a +use_puppet_ca_cert+ flag in your beaker host.cfg. If present,
the beaker-pe install_pe() method will first scp the master's ca.pem to
the agent, and then curl the frictionless installer script with
--cacert.

The install_ca_cert() method is being called separately from
installer_cmd() because the later just returns the install command
string to be executed on a host, and I did not want to embed
a side effect copying files inside installer_cmd().

Note that the current implementation only addresses *nix platforms
(except OSX). Windows and OSX are to be addressed in PE-21368.