Added log message to long running section
Right before a long pause in the install run, was a message about scp'ing a file.  Clearly that wasn't taking all the time.
Turns out that pe was being installed on the master right after that, but no message was logged.
So added a message to log that fact and make it clear.