(maint) Allow unauthenticated packages on debian based systems
Previously `install_dev_repos_on` only installed the dev repos on a
system.

Unfortunately, the dev repos are unsigned and apt-get
installation will fail for packages provided by the dev repo. The
console services all have copied this method and updated it to allow
unauthenticated packages. Doing so is required for a clean PE
installation using the dev repos.

This patch updates the upstream method in this repo to be functionally
equivalent to the copied methods used in downstream consumers, allowing
those downstream consumers to no longer maintain copy-pasted versions of
this method.